@@ -5,13 +5,24 @@ import {ScopeError} from "../ContextSensitiveError";
 import {AppliedArgumentToVariable} from "../ContextSensitiveError";
 
 export class ErrorOutputting extends SuperVisitor {
-    foundErrors = false;
+    print = true;
+    errors: ContextSensitiveError[]  = [];
+
+    constructor(print?: boolean) {
+        this.print = print;
+    }
+
+    get foundErrors() {
+        return this.errors.length > 0;
+    }
 
     afterVisitNode(node: A.AbstractSyntaxType) {
         if (node.errors && node.errors.length > 0) {
-            this.foundErrors = true;
-            for (let error of node.errors) {
-                console.error(error.toString());
+            this.errors = this.errors.concat(node.errors);
+            if (print) {
+                for (let error of node.errors) {
+                    console.error(error.toString());
+                }
             }
         }
     }
